// services/execd/tests/generated/v1/pkgd.ts
export enum InterfaceProtocol {
  INTERFACE_PROTOCOL_UNSPECIFIED = 0,
  INTERFACE_PROTOCOL_HTTP = 1,
  INTERFACE_PROTOCOL_GRPC = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface App {
  appId: string;
  scope: AppScope | undefined;
  placementId: string;
  packageId: string;
  desiredPackageGeneration: bigint;
  revision: bigint;
  createdAt: Date | undefined;
  updatedAt: Date | undefined;
}

export interface AppScope {
  global?: GlobalAppScope | undefined;
  tenant?: TenantAppScope | undefined;
  workspace?: WorkspaceAppScope | undefined;
  user?: UserAppScope | undefined;
}

export interface GlobalAppScope {
}

export interface TenantAppScope {
  tenantId: string;
}

export interface WorkspaceAppScope {
  tenantId: string;
  workspaceId: string;
}
// ...
export const AppScope: MessageFns<AppScope> = {
  encode(message: AppScope, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.global !== undefined) {
      GlobalAppScope.encode(message.global, writer.uint32(10).fork()).join();
    }
    if (message.tenant !== undefined) {
      TenantAppScope.encode(message.tenant, writer.uint32(18).fork()).join();
    }
    if (message.workspace !== undefined) {
      WorkspaceAppScope.encode(message.workspace, writer.uint32(26).fork()).join();
    }
    if (message.user !== undefined) {
      UserAppScope.encode(message.user, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppScope {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppScope();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.global = GlobalAppScope.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.tenant = TenantAppScope.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.workspace = WorkspaceAppScope.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.user = UserAppScope.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AppScope {
    return {
      global: isSet(object.global) ? GlobalAppScope.fromJSON(object.global) : undefined,
      tenant: isSet(object.tenant) ? TenantAppScope.fromJSON(object.tenant) : undefined,
      workspace: isSet(object.workspace) ? WorkspaceAppScope.fromJSON(object.workspace) : undefined,
      user: isSet(object.user) ? UserAppScope.fromJSON(object.user) : undefined,
    };
  },

  toJSON(message: AppScope): unknown {
    const obj: any = {};
    if (message.global !== undefined) {
      obj.global = GlobalAppScope.toJSON(message.global);
    }
    if (message.tenant !== undefined) {
      obj.tenant = TenantAppScope.toJSON(message.tenant);
    }
    if (message.workspace !== undefined) {
      obj.workspace = WorkspaceAppScope.toJSON(message.workspace);
    }
    if (message.user !== undefined) {
      obj.user = UserAppScope.toJSON(message.user);
    }
    return obj;
  },

  create(base?: DeepPartial<AppScope>): AppScope {
    return AppScope.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<AppScope>): AppScope {
    const message = createBaseAppScope();
    message.global = (object.global !== undefined && object.global !== null)
      ? GlobalAppScope.fromPartial(object.global)
      : undefined;
    message.tenant = (object.tenant !== undefined && object.tenant !== null)
      ? TenantAppScope.fromPartial(object.tenant)
      : undefined;
    message.workspace = (object.workspace !== undefined && object.workspace !== null)
      ? WorkspaceAppScope.fromPartial(object.workspace)
      : undefined;
    message.user = (object.user !== undefined && object.user !== null)
      ? UserAppScope.fromPartial(object.user)
      : undefined;
    return message;
  },
};

function createBaseGlobalAppScope(): GlobalAppScope {
  return {};
}

export const GlobalAppScope: MessageFns<GlobalAppScope> = {
  encode(_: GlobalAppScope, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GlobalAppScope {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGlobalAppScope();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): GlobalAppScope {
    return {};
  },

  toJSON(_: GlobalAppScope): unknown {
    const obj: any = {};
    return obj;
  },

  create(base?: DeepPartial<GlobalAppScope>): GlobalAppScope {
    return GlobalAppScope.fromPartial(base ?? {});
  },
  fromPartial(_: DeepPartial<GlobalAppScope>): GlobalAppScope {
    const message = createBaseGlobalAppScope();
    return message;
  },
};

function createBaseTenantAppScope(): TenantAppScope {
  return { tenantId: "" };
}

export const TenantAppScope: MessageFns<TenantAppScope> = {
  encode(message: TenantAppScope, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.tenantId !== "") {
      writer.uint32(10).string(message.tenantId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TenantAppScope {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTenantAppScope();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.tenantId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TenantAppScope {
    return { tenantId: isSet(object.tenantId) ? globalThis.String(object.tenantId) : "" };
  },

  toJSON(message: TenantAppScope): unknown {
    const obj: any = {};
    if (message.tenantId !== "") {
      obj.tenantId = message.tenantId;
    }
    return obj;
  },

  create(base?: DeepPartial<TenantAppScope>): TenantAppScope {
    return TenantAppScope.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TenantAppScope>): TenantAppScope {
    const message = createBaseTenantAppScope();
    message.tenantId = object.tenantId ?? "";
    return message;
  },
};

function createBaseWorkspaceAppScope(): WorkspaceAppScope {
  return { tenantId: "", workspaceId: "" };
}

export const WorkspaceAppScope: MessageFns<WorkspaceAppScope> = {
  encode(message: WorkspaceAppScope, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.tenantId !== "") {
      writer.uint32(10).string(message.tenantId);
    }
    if (message.workspaceId !== "") {
      writer.uint32(18).string(message.workspaceId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): WorkspaceAppScope {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseWorkspaceAppScope();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.tenantId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.workspaceId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): WorkspaceAppScope {
    return {
      tenantId: isSet(object.tenantId) ? globalThis.String(object.tenantId) : "",
      workspaceId: isSet(object.workspaceId) ? globalThis.String(object.workspaceId) : "",
    };
  },

  toJSON(message: WorkspaceAppScope): unknown {
    const obj: any = {};
    if (message.tenantId !== "") {
      obj.tenantId = message.tenantId;
    }
    if (message.workspaceId !== "") {
      obj.workspaceId = message.workspaceId;
    }
    return obj;
  },

  create(base?: DeepPartial<WorkspaceAppScope>): WorkspaceAppScope {
    return WorkspaceAppScope.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<WorkspaceAppScope>): WorkspaceAppScope {
    const message = createBaseWorkspaceAppScope();
    message.tenantId = object.tenantId ?? "";
    message.workspaceId = object.workspaceId ?? "";
    return message;
  },
};
// ...
export interface PackageServiceClient extends Client {
  declarePackage(
    request: DeclarePackageRequest,
    callback: (error: ServiceError | null, response: Package) => void,
  ): ClientUnaryCall;
  declarePackage(
    request: DeclarePackageRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: Package) => void,
  ): ClientUnaryCall;
  declarePackage(
    request: DeclarePackageRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: Package) => void,
  ): ClientUnaryCall;
  getPackage(
    request: GetPackageRequest,
    callback: (error: ServiceError | null, response: Package) => void,
  ): ClientUnaryCall;
  getPackage(
    request: GetPackageRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: Package) => void,
  ): ClientUnaryCall;
  getPackage(
    request: GetPackageRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: Package) => void,
  ): ClientUnaryCall;
  createApp(request: CreateAppRequest, callback: (error: ServiceError | null, response: App) => void): ClientUnaryCall;
  createApp(
    request: CreateAppRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: App) => void,
  ): ClientUnaryCall;
  createApp(
    request: CreateAppRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: App) => void,
  ): ClientUnaryCall;
  getApp(request: GetAppRequest, callback: (error: ServiceError | null, response: App) => void): ClientUnaryCall;
  getApp(
    request: GetAppRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: App) => void,
  ): ClientUnaryCall;
  getApp(
    request: GetAppRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: App) => void,
  ): ClientUnaryCall;
  setAppPackageGeneration(
    request: SetAppPackageGenerationRequest,
    callback: (error: ServiceError | null, response: App) => void,
  ): ClientUnaryCall;
  setAppPackageGeneration(
    request: SetAppPackageGenerationRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: App) => void,
  ): ClientUnaryCall;
  setAppPackageGeneration(
    request: SetAppPackageGenerationRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: App) => void,
  ): ClientUnaryCall;
}

export const PackageServiceClient = makeGenericClientConstructor(
  PackageServiceService,
  "ctlflow.packages.v1.PackageService",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): PackageServiceClient;
  service: typeof PackageServiceService;
  serviceName: string;
};